Temp build to work on a different device

+++ src/events/generalevents.ts
@@ -1,7 +1,7 @@
 import Discord from 'discord.js'
 import {client, blackjackSessions} from '../main';
-import {prefix,session_category_id} from '../config.json';
-import {queryPromise} from '../utils';
+import {session_category_id} from '../config.json';
+import {queryPromise, getGuildPrefix} from '../utils';
 import { blacklistedChannels } from '../staticdata';
 
 export function SetupEvents()
@@ -26,13 +26,23 @@ async function onMSGReceived(msg: Discord.Message)
         if (msg.author.bot) return;
         //relay channel messages when blackjack session is active
         if ((msg.channel as Discord.GuildChannel).parentID == session_category_id && blackjackSessions.find(x => x.user.id == msg.author.id)) return blackjackSessions.find(x => x.user.id == msg.author.id)!.handleSessionCommand(msg);
-    
+        
+        //Split args
+        const args: string[] = msg.content.split(/ +/);
+        var command = args.shift()!.toLowerCase();
+
+        //check if we're running the prefix changing command
+        if (command == "rpgthunder") {changePrefixCommand(msg,args); return;}
+
+        //Get prefix
+        const prefix = await getGuildPrefix(msg.guild.id);
+        command = command.slice(prefix.length);
+
         //Check if it starts with required refix
         if (!msg.content.startsWith(prefix)) return;
+
         if (blacklistedChannels.includes(msg.channel.id)) {await msg.delete(); return;}
-        //Split args and execute command if it exists.
-        const args: string[] = msg.content.slice(prefix.length).split(/ +/);
-        const command = args.shift()!.toLowerCase();
+        //Execute command if it exists.
         if (client.commands.has(command))  client.commands.get(command).execute(msg, args);
         else if (client.commands.find((x:any) => x.aliases.find((alias:string) => alias == command))) client.commands.find(x => x.aliases.find((alias:string) => alias == command)).execute(msg,args);
     }
@@ -44,10 +54,38 @@ async function onMSGReceived(msg: Discord.Message)
     
 }
 
+async function changePrefixCommand(msg:Discord.Message, args:string[])
+{
+    try
+    {
+        switch(args[0].toLowerCase())
+        {
+            case "setprefix":
+                if (!msg.member.permissions.has("ADMINISTRATOR")) return msg.reply("You do not have permission to execute that command.");
+                if(args[1].length == 0) msg.reply("Please enter what you would like to change the prefix to.\nUsage: `rpgthunder setprefix [prefix]`");
+            
+                if (args[1].length > 10) return msg.reply("Your custom prefix cannot be longer than 10 characters.");
+            
+                //check if it already exists
+                const result = (await queryPromise(`SELECT * FROM custom_prefix WHERE guild_id=${msg.guild.id}`))[0];
+
+                if(result) await queryPromise(`UPDATE custom_prefix SET prefix='${args[1]}' WHERE guild_id=${msg.guild.id}`);
+                else await queryPromise(`INSERT INTO custom_prefix(guild_id,prefix) VALUES('${msg.guild.id}','${args[1]}')`);
+            
+                msg.reply(`Prefix has been changed to \`${args[1]}\`.`);
+                break;
+            case "prefix":
+                msg.reply(`This guilds' prefix is \`${await getGuildPrefix(msg.guild.id)}\``);
+        }
+    }
+    catch(err){console.log(err);}
+   
+   
+}
 
 //Timed events.
 async function updateBotStatus(){
     var countResult = await queryPromise("SELECT COUNT(*) from users")
     var registerCount = countResult[0][Object.keys(countResult[0])[0]]
-    client.c.user.setActivity(`${prefix}help | ${registerCount} registered users on ${client.c.guilds.size} servers`);
+    client.c.user.setActivity(`$help | ${registerCount} registered users on ${client.c.guilds.size} servers`);
 }
